Made the date range form its own component
